MyAppGlidModuleは別個のクラスファイルにした方が良い
